+++ clang-tools-extra/clang-tidy/hooks/EntryPointRecursionCheck.cpp
@@ -1,0 +1,30 @@
+//===--- EntryPointRecursionCheck.cpp - clang-tidy ------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "EntryPointRecursionCheck.h"
+#include "clang/AST/ASTContext.h"
+#include "clang/ASTMatchers/ASTMatchFinder.h"
+
+using namespace clang::ast_matchers;
+
+namespace clang {
+namespace tidy {
+namespace hooks {
+
+void EntryPointRecursionCheck::registerMatchers(MatchFinder *Finder) {
+  Finder->addMatcher(callExpr(callee(functionDecl(anyOf(hasName("cbak"), hasName("hook"))))).bind("recursion"), this);
+}
+
+void EntryPointRecursionCheck::check(const MatchFinder::MatchResult &Result) {
+  const auto *Matched = Result.Nodes.getNodeAs<CallExpr>("recursion");
+  diag(Matched->getExprLoc(), "hook functions may not call each other (nor themselves)");
+}
+
+} // namespace hooks
+} // namespace tidy
+} // namespace clang