+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -101,6 +101,11 @@ New checks
   Finds virtual classes whose destructor is neither public and virtual nor
   protected and non-virtual.
 
+- New :doc:`hooks-entry-point-recursion
+  <clang-tidy/checks/hooks-entry-point-recursion>` check.
+
+  FIXME: add release notes.
+
 - New :doc:`hooks-guard-called
   <clang-tidy/checks/hooks-guard-called>` check.